Update vendoring mechanism to use pip-tools

+++ tools/vendored.py
@@ -13,79 +13,57 @@ def remove_all(paths):
 
 
 def update_vendored():
-    update_pkg_resources()
-    update_setuptools()
+    upgrade = "--upgrade" in sys.argv  # Very simple, no need for argparse yet.
+    update_target('pkg_resources', upgrade=upgrade)
+    update_target(
+        'setuptools',
+        upgrade=upgrade,
+        constraints=Path('pkg_resources/_vendor/vendored.txt'),  # ensure compatibility
+    )
 
 
-def rewrite_packaging(pkg_files, new_root):
+def repair_imports(vendor):
     """
-    Rewrite imports in packaging to redirect to vendored copies.
+    Rewrite imports to redirect to vendored copies.
+    This applies to all Python files inside the ``_vendor`` directory.
     """
-    for file in pkg_files.glob('*.py'):
-        text = file.text()
-        text = re.sub(r' (pyparsing)', rf' {new_root}.\1', text)
-        text = text.replace(
-            'from six.moves.urllib import parse',
-            'from urllib import parse',
-        )
-        file.write_text(text)
+    target = vendor.parent.name  # setuptools or pkg_resources
+    names = rf"(?P<name>{'|'.join(yield_top_level(target))})"
 
-
-def rewrite_jaraco_text(pkg_files, new_root):
-    """
-    Rewrite imports in jaraco.text to redirect to vendored copies.
-    """
-    for file in pkg_files.glob('*.py'):
+    for file in vendor.walkfiles(match="*.py"):
         text = file.read_text()
-        text = re.sub(r' (jaraco\.)', rf' {new_root}.\1', text)
-        text = re.sub(r' (importlib_resources)', rf' {new_root}.\1', text)
-        # suppress loading of lorem_ipsum; ref #3072
-        text = re.sub(r'^lorem_ipsum.*\n$', '', text, flags=re.M)
+        text = re.sub(  # import {name} => from setupools.extern import {name}
+            rf"^(?P<indent>\s*)import\s+{names}(?P<remaining>.*)$",
+            rf'\g<indent>from {target}.extern import \g<name>\g<remaining>',
+            # ^--- Assumes {name} is not nested.
+            #      Otherwise we need to pass `repl=<function>` to split the import
+            #      in multiple lines (one import for the parent and one for the child).
+            text,
+            flags=re.MULTILINE,
+        )
+        text = re.sub(  # from {name} => from setuptools.extern.{name}
+            rf"^(?P<indent>\s*)from\s+{names}(?P<remaining>.* import .*)$",
+            rf'\g<indent>from {target}.extern.\g<name>\g<remaining>',
+            text,
+            flags=re.MULTILINE,
+        )
         file.write_text(text)
 
 
 def repair_namespace(pkg_files):
     # required for zip-packaged setuptools #3084
-    pkg_files.joinpath('__init__.py').write_text('')
+    if pkg_files.is_dir() and not (pkg_files / "__init__.py").exists():
+        (pkg_files / '__init__.py').write_text('')
 
 
-def rewrite_jaraco_functools(pkg_files, new_root):
+def rewrite_jaraco_text(pkg_files):
     """
-    Rewrite imports in jaraco.functools to redirect to vendored copies.
-    """
-    for file in pkg_files.glob('*.py'):
-        text = file.read_text()
-        text = re.sub(r' (more_itertools)', rf' {new_root}.\1', text)
-        file.write_text(text)
-
-
-def rewrite_jaraco_context(pkg_files, new_root):
-    """
-    Rewrite imports in jaraco.context to redirect to vendored copies.
-    """
-    for file in pkg_files.glob('context.py'):
-        text = file.read_text()
-        text = re.sub(r' (backports)', rf' {new_root}.\1', text)
-        file.write_text(text)
-
-
-def rewrite_importlib_resources(pkg_files, new_root):
-    """
-    Rewrite imports in importlib_resources to redirect to vendored copies.
-    """
-    for file in pkg_files.glob('*.py'):
-        text = file.read_text().replace('importlib_resources.abc', '.abc')
-        text = text.replace('zipp', '..zipp')
-        file.write_text(text)
-
-
-def rewrite_importlib_metadata(pkg_files, new_root):
-    """
-    Rewrite imports in importlib_metadata to redirect to vendored copies.
+    Rewrite imports in jaraco.text to redirect to vendored copies.
     """
     for file in pkg_files.glob('*.py'):
         text = file.read_text()
-        text = text.replace('import zipp', 'from .. import zipp')
+        # suppress loading of lorem_ipsum; ref #3072
+        text = re.sub(r'^lorem_ipsum.*\n$', '', text, flags=re.M)
         file.write_text(text)
 
 
@@ -94,6 +72,8 @@ def rewrite_more_itertools(pkg_files: Path):
     Defer import of concurrent.futures. Workaround for #3090.
     """
     more_file = pkg_files.joinpath('more.py')
+    if not more_file.exists():
+        return
     text = more_file.read_text()
     text = re.sub(r'^.*concurrent.futures.*?\n', '', text, flags=re.MULTILINE)
     text = re.sub(
@@ -109,6 +89,9 @@ def rewrite_wheel(pkg_files: Path):
     Remove parts of wheel not needed by bdist_wheel, and rewrite imports to use
     setuptools's own code or vendored dependencies.
     """
+    if not pkg_files.exist():
+        return
+
     shutil.rmtree(pkg_files / 'cli')
     shutil.rmtree(pkg_files / 'vendored')
     pkg_files.joinpath('_setuptools_logging.py').unlink()
@@ -158,34 +141,46 @@ def rewrite_wheel(pkg_files: Path):
             path.write_text(code)  # type: ignore[attr-defined]
 
 
-def rewrite_platformdirs(pkg_files: Path):
-    """
-    Replace some absolute imports with relative ones.
-    """
-    init = pkg_files.joinpath('__init__.py')
-    text = init.read_text()
-    text = text.replace('from platformdirs.', 'from .')
-    init.write_text(text)
-
-
 def clean(vendor):
     """
     Remove all files out of the vendor directory except the meta
     data (as pip uninstall doesn't support -t).
     """
-    ignored = ['vendored.txt', 'ruff.toml']
-    remove_all(path for path in vendor.glob('*') if path.basename() not in ignored)
+    ignored = ['vendored.in', 'vendored.txt', 'ruff.toml']
+    remove_all(path for path in vendor.glob('*') if path.name not in ignored)
 
 
-def install(vendor):
-    clean(vendor)
+def resolve(vendor, upgrade=False, constraints=None):
+    """Document why each indirect dependency is required."""
+    constraint_args = ["--constraint", str(constraints)] if constraints else []
+    upgrade_args = ["--upgrade"] if upgrade else []
+    args = [
+        sys.executable,
+        '-m',
+        'piptools',
+        'compile',
+        '--strip-extras',
+        *upgrade_args,
+        *constraint_args,
+        '-o',
+        str(vendor / 'vendored.txt'),
+        str(vendor / 'vendored.in'),
+    ]
+    subprocess.check_call(args)
+
+
+def install(vendor, constraints=None):
+    constraint_args = ["--constraint", str(constraints)] if constraints else []
     install_args = [
         sys.executable,
         '-m',
         'pip',
         'install',
-        '-r',
+        *constraint_args,
+        '-c',
         str(vendor / 'vendored.txt'),
+        '-r',
+        str(vendor / 'vendored.in'),
         '-t',
         str(vendor),
     ]
@@ -193,33 +188,20 @@ def install(vendor):
     (vendor / '__init__.py').write_text('')
 
 
-def update_pkg_resources():
-    vendor = Path('pkg_resources/_vendor')
-    install(vendor)
-    rewrite_packaging(vendor / 'packaging', 'pkg_resources.extern')
+def update_target(target, upgrade=None, constraints=None):
+    vendor = Path(f'{target}/_vendor')
+    clean(vendor)
+    resolve(vendor, upgrade, constraints)
+    install(vendor, constraints)
+
+    # Patches related to issues
     repair_namespace(vendor / 'jaraco')
     repair_namespace(vendor / 'backports')
-    rewrite_jaraco_text(vendor / 'jaraco/text', 'pkg_resources.extern')
-    rewrite_jaraco_functools(vendor / 'jaraco/functools', 'pkg_resources.extern')
-    rewrite_jaraco_context(vendor / 'jaraco', 'pkg_resources.extern')
-    rewrite_importlib_resources(vendor / 'importlib_resources', 'pkg_resources.extern')
-    rewrite_more_itertools(vendor / "more_itertools")
-    rewrite_platformdirs(vendor / "platformdirs")
-
-
-def update_setuptools():
-    vendor = Path('setuptools/_vendor')
-    install(vendor)
-    rewrite_packaging(vendor / 'packaging', 'setuptools.extern')
-    repair_namespace(vendor / 'jaraco')
-    repair_namespace(vendor / 'backports')
-    rewrite_jaraco_text(vendor / 'jaraco/text', 'setuptools.extern')
-    rewrite_jaraco_functools(vendor / 'jaraco/functools', 'setuptools.extern')
-    rewrite_jaraco_context(vendor / 'jaraco', 'setuptools.extern')
-    rewrite_importlib_resources(vendor / 'importlib_resources', 'setuptools.extern')
-    rewrite_importlib_metadata(vendor / 'importlib_metadata', 'setuptools.extern')
+    rewrite_jaraco_text(vendor / 'jaraco/text')
     rewrite_more_itertools(vendor / "more_itertools")
     rewrite_wheel(vendor / "wheel")
+
+    repair_imports(vendor)  # Required for a vendoring workflow
 
 
 def yield_top_level(name):
